add index and cluster statistics
This commit adds index and cluster statistics. A big portion of the code
was forked from botify-labs/collectd-elasticsearch. The index name is
encoded using [] which would make it appear as a dimension when reported
to a SignalFx endpoint.

This commit cleans up the code to make it a bit more pep8 compliant. It
also adds collectd mocking to make this easy to test by just running the
python script.