and more counters + consistency/types cleanups
* add per index merge, filter cache and field data statistics
* make collectd metric names consistent with the counters that have been
there when we forked the repo (replace '_' with '-' and drop units). I'm
not too fond of doing this (I would have preferred sticking with
elasticsearch's naming), but it makes finding those metrics consistent
regardless of whether you are looking for per node, per index (primaries
vs total) or per cluster statistics.
* update bytes to be gauges since bytes are sent as counters by collectd
and most of the time it's just a gauge. There are a couple of bytes that
I left unchanged because those were really counters.