Removing legacy files
The install.sh script in this repo is over 2 years old and is completely
invalid and unnecessary. The elasticsearch.conf file is an out-of-date
subset of the real one which is located at
https://github.com/signalfx/integrations/blob/master/collectd-elasticsearch/20-elasticsearch.conf.
Removing both of these files from this repo as their continued presence
is unnecessary and is causing confusion.